refactor slash command stuff out of index
